simplify networking
