+++ test/network-request-test.js
@@ -4,7 +4,7 @@ import util from './util';
 
 global.window = util.setupWindow();
 
-const NetworkService = require('../extension/src/network-service.js');
+const networkRequest = require('../extension/src/network-request.js');
 
 test.beforeEach(t => {
 	t.context.endpoint = 'http://endpoint.net/foo';
@@ -14,7 +14,7 @@ test('#request returns Promise', t => {
 	window.fetch = sinon.stub().returns(Promise.resolve('{}'));
 	window.localStorage.getItem = sinon.stub().returns('oauthToken');
 
-	return NetworkService.request(t.context.endpoint).then(() => {
+	return networkRequest(t.context.endpoint).then(() => {
 		t.pass();
 	});
 });
@@ -23,7 +23,7 @@ test('#request requests fetches given url with proper headers', t => {
 	window.fetch = sinon.stub().returns(Promise.resolve('{}'));
 	window.localStorage.getItem = sinon.stub().returns('oauthToken');
 
-	return NetworkService.request(t.context.endpoint).then(() => {
+	return networkRequest(t.context.endpoint).then(() => {
 		const args = [t.context.endpoint, {
 			headers: {
 				'Authorization': 'token oauthToken',
@@ -37,5 +37,5 @@ test('#request requests fetches given url with proper headers', t => {
 test('#request returns rejected Promise if oauthToken is empty', t => {
 	window.localStorage.getItem = sinon.stub().returns('');
 
-	t.throws(NetworkService.request(t.context.endpoint), 'missing token');
+	t.throws(networkRequest(t.context.endpoint), 'missing token');
 });
